Use feign client on ShoppingCartService

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interaction/api/commerce/dto/warehouse/ShoppingCartDto.java
@@ -17,5 +17,5 @@ public class ShoppingCartDto {
     private UUID shoppingCartId;
 
     @NotNull
-    private Map<UUID, Integer> products;
+    private Map<UUID, Long> products;
 }
